+++ tests/testthat/scripts/todos.R
@@ -1,0 +1,9 @@
+# This is a file that doesn't do much
+
+# TODO make x longer
+x <- 1:10
+length(x)
+
+# TODO add another example
+
+# FIXME This is a fixme